Add history board
