+++ src/app/utils/index.ts
@@ -1,4 +1,5 @@
-import { IdentifiableQuestion } from "@interfaces/type";
+import { QuestionState } from "@interfaces/db";
+import { IdentifiableQuestion, IdentifiableQuestions } from "@interfaces/type";
 import { FieldValue, Timestamp } from "firebase/firestore";
 
 export const trimName = (name: string) => {
@@ -30,9 +31,13 @@ export const compareQuestions = (
 };
 
 export const formatTimeDifference = (
-  timestamp: FieldValue | Timestamp
+  question: IdentifiableQuestion
 ): string => {
-  const date = timestamp instanceof Timestamp ? timestamp.toDate() : new Date();
+  const date = question.timestamp.toDate();
+
+  if (hasPassed(question)) {
+    return date.toLocaleDateString();
+  }
 
   const now = new Date();
   const diffInSeconds = Math.floor((now.getTime() - date.getTime()) / 1000);
@@ -57,3 +62,18 @@ export const formatTimeDifference = (
 
   return `${formattedHours}:${formattedMinutes} ${period}`;
 };
+
+/**
+ * A question has "expired" if it's a different date or it's been resolved
+ */
+export const hasPassed = (question: IdentifiableQuestion) => {
+  const postedAt = question.timestamp.toDate().setHours(0, 0, 0, 0);
+  const now = new Date().setHours(0, 0, 0, 0);
+  return now > postedAt || question.state === QuestionState.RESOLVED;
+};
+
+export const getActiveQuestions = (questions: IdentifiableQuestions) =>
+  questions.filter((question) => !hasPassed(question));
+
+export const getExpiredQuestions = (questions: IdentifiableQuestions) =>
+  questions.filter((question) => hasPassed(question));
